infra/functions/synthetic_health_check/requirements.txt
requests
google-auth
google-auth-httplib2
google-cloud-monitoring
